+++ Assets/Animation/Flight/EnemyOrigin/TakeDamage.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: af81f25588cfed24ea46f25a99f19976
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
