+++ Assets/Animation/Flight/TakeDamage.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 50208ba966a3a244694db8887fcc54a8
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
