+++ Assets/Animation/Flight/TakeDamage/PlayerInvisibility.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 754a0a3a1d7570d4fb0294ecea9d7652
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
