+++ Assets/Fonts/DungGeunMo SDF.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8935ba29fc1ab9c40a1fe4c33b764572
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
