+++ Assets/Prefabs/InGamePrefabs/Player_Update.prefab
@@ -283,7 +283,7 @@ MonoBehaviour:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 87775265348895361}
-  m_Enabled: 1
+  m_Enabled: 0
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 3771a69bb2511b5489864299c6be4ec2, type: 3}
   m_Name: 
@@ -1487,7 +1487,7 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 7332932538739181366, guid: d68ab9a631ce844419d459f851941164, type: 3}
       propertyPath: m_LocalScale.y
-      value: 11
+      value: 16.05265
       objectReference: {fileID: 0}
     - target: {fileID: 7332932538739181366, guid: d68ab9a631ce844419d459f851941164, type: 3}
       propertyPath: m_LocalPosition.x
@@ -1495,7 +1495,7 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 7332932538739181366, guid: d68ab9a631ce844419d459f851941164, type: 3}
       propertyPath: m_LocalPosition.y
-      value: 4.62
+      value: 6.8656
       objectReference: {fileID: 0}
     - target: {fileID: 7332932538739181366, guid: d68ab9a631ce844419d459f851941164, type: 3}
       propertyPath: m_LocalPosition.z
