--- Assets/Scripts/GameData/EnemyFlightData.cs
@@ -1,8 +1,0 @@
-﻿using UnityEngine;
-
-public class EnemyFlightData : FlightData
-{
-    [Header("# Enemy Stat")]  
-    [SerializeField] private EEnemyType E_enemyType;
-    
-}
