--- Assets/Scripts/GameData/EnemyFlightData.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 4b472ff3b6a640bab49c267262d1d2d4
-timeCreated: 1715831793
