--- Assets/Scripts/GameData/FlightData.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 45339c5cc5f70724baccf9b20dbb4ee5
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
