+++ Assets/Scripts/GamePlay/Controllers/AnimationControllers/FlightAnimationController.cs.meta
@@ -1,3 +1,11 @@
 fileFormatVersion: 2
-guid: 5b150f71e8a34b01ad894948ee585f86
-timeCreated: 1715942589
+guid: 1290508b2c8a1544b844d6ca2a607ab0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
