+++ Assets/Scripts/GamePlay/Controllers/AnimationControllers/PlayerAnimationController.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 5b150f71e8a34b01ad894948ee585f86
+timeCreated: 1715942589
