--- Assets/Scripts/GamePlay/Controllers/AnimationControllers/TopDownAnimationController.cs
@@ -1,13 +1,0 @@
-using UnityEngine;
-
-public class TopDownAnimationController : MonoBehaviour
-{
-    protected TopDownController topDownController;
-    protected Animator anim;
-
-    protected virtual void Awake()
-    {
-        topDownController = GetComponent<TopDownController>();
-        anim = GetComponent<Animator>();
-    }
-}
