--- Assets/Scripts/GamePlay/Controllers/AnimationControllers/TopDownAnimationController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1290508b2c8a1544b844d6ca2a607ab0
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
