+++ Assets/Scripts/GamePlay/Stats/FlightStat.cs
@@ -7,7 +7,7 @@ public class FlightStat
     [Header("# Flight Stat")]
     public float MoveSpeed;
     public float AtkDelay;
-    [Range(1f,1f)]public float AtkDamage;
+    [Range(1f,5f)]public float AtkDamage;
     [Range(0f, 100f)]public float MaxHealth;
     [Range(1, 100)] public int BulletNum;
     [Range(0f, 360f)] public float BulletAngle;
