+++ ProjectSettings/TagManager.asset
@@ -17,7 +17,7 @@ TagManager:
   - Player
   - Enemy
   - Lazer
-  - 
+  - PlayerDamaged
   - 
   - 
   -
